+++ tests/test_processing.py
@@ -950,6 +950,107 @@ class TestSliceExtraction(unittest.TestCase):
             self.dataset.process(self.processing)
 
 
+class TestSliceRemoval(unittest.TestCase):
+    def setUp(self):
+        self.processing = aspecd.processing.SliceRemoval()
+        self.dataset = aspecd.dataset.Dataset()
+        data = np.sin(np.linspace(0, 2*np.pi, num=500))
+        self.dataset.data.data = np.tile(data, (5, 1)) \
+            * np.tile(np.linspace(1, 2, num=5), (500, 1)).T
+
+    def test_instantiate_class(self):
+        pass
+
+    def test_has_appropriate_description(self):
+        self.assertIn('remove slice', self.processing.description.lower())
+
+    def test_is_undoable(self):
+        self.assertTrue(self.processing.undoable)
+
+    def test_with_1D_dataset_raises(self):
+        dataset = aspecd.dataset.Dataset()
+        dataset.data.data = np.sin(np.linspace(0, 2*np.pi, num=500))
+        with self.assertRaises(
+                aspecd.exceptions.NotApplicableToDatasetError):
+            dataset.process(self.processing)
+
+    def test_without_position_raises(self):
+        with self.assertRaisesRegex(IndexError, "for slice extraction"):
+            self.dataset.process(self.processing)
+
+    def test_with_index_exceeding_dimension_raises(self):
+        self.processing.parameters['position'] = 10
+        with self.assertRaisesRegex(ValueError, "Position out of axis range."):
+            self.dataset.process(self.processing)
+
+    def test_remove_slice(self):
+        origdata = self.dataset.data.data
+        self.processing.parameters['position'] = 3
+        self.dataset.process(self.processing)
+        # This assertion is currently wrong ([:2, :])
+        np.testing.assert_allclose(np.delete(origdata, 3, 0),
+                                   self.dataset.data.data)
+
+    def test_remove_slice_with_index_zero(self):
+        origdata = self.dataset.data.data
+        self.processing.parameters['position'] = 0
+        self.dataset.process(self.processing)
+        np.testing.assert_allclose(origdata[1:, :], self.dataset.data.data)
+
+    def test_remove_slice_operates_along_first_axis_by_default(self):
+        origdata = self.dataset.data.data
+        self.processing.parameters['position'] = 3
+        self.dataset.process(self.processing)
+        np.testing.assert_allclose(np.delete(origdata, 3, 0),
+                                   self.dataset.data.data)
+
+    def test_remove_slice_along_second_axis(self):
+        origdata = self.dataset.data.data
+        self.processing.parameters['position'] = 3
+        self.processing.parameters['axis'] = 1
+        self.dataset.process(self.processing)
+        np.testing.assert_allclose(np.delete(origdata, 3, 1),
+                                   self.dataset.data.data)
+
+    def test_remove_slice_along_non_existing_axis_raises(self):
+        with self.assertRaisesRegex(IndexError, "Axis [0-9]+ out of bounds"):
+            self.processing.parameters['position'] = 3
+            self.processing.parameters['axis'] = 2
+            self.dataset.process(self.processing)
+
+    def test_remove_slice_with_wrong_unit_raises(self):
+        self.processing.parameters["unit"] = "foo"
+        self.processing.parameters["position"] = 3
+        with self.assertRaises(ValueError):
+            self.dataset.process(self.processing)
+
+    def test_remove_slice_with_axis_units(self):
+        self.processing.parameters["unit"] = "axis"
+        self.dataset.data.axes[0].values = \
+            np.linspace(30, 70, len(self.dataset.data.axes[0].values))
+        self.processing.parameters["position"] = 40
+        data = np.delete(self.dataset.data.data, 1, 0)
+        self.dataset.process(self.processing)
+        np.testing.assert_allclose(data, self.dataset.data.data)
+
+    def test_unit_is_case_insensitive(self):
+        self.processing.parameters["unit"] = "aXis"
+        self.dataset.data.axes[0].values = \
+            np.linspace(30, 70, len(self.dataset.data.axes[0].values))
+        self.processing.parameters["position"] = 40
+        data = np.delete(self.dataset.data.data, 1, 0)
+        self.dataset.process(self.processing)
+        np.testing.assert_allclose(data, self.dataset.data.data)
+
+    def test_remove_slice_with_axis_units_out_of_range_raises(self):
+        self.processing.parameters["unit"] = "axis"
+        self.dataset.data.axes[0].values = \
+            np.linspace(30, 70, len(self.dataset.data.axes[0].values))
+        self.processing.parameters["position"] = 300
+        with self.assertRaises(ValueError):
+            self.dataset.process(self.processing)
+
+
 class TestRangeExtraction(unittest.TestCase):
     def setUp(self):
         self.processing = aspecd.processing.RangeExtraction()